requirements.txt
asgiref==3.8.1
bleach==6.2.0
chardet==5.2.0
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
reportlab==4.2.5
sqlparse==0.5.3
webencodings==0.5.1
